refactor: extract conversation navigation results
